Rename files to snake case. Fix params for augmentation().
